+++ sdk/dotnet/Remote/Outputs/Connection.cs
@@ -39,7 +39,7 @@ namespace Pulumi.File.Remote.Outputs
         /// <summary>
         /// The port to connect to.
         /// </summary>
-        public readonly double? Port;
+        public readonly int? Port;
         /// <summary>
         /// The contents of an SSH key to use for the connection. This takes preference over the password if provided.
         /// </summary>
@@ -69,7 +69,7 @@ namespace Pulumi.File.Remote.Outputs
 
             int? perDialTimeout,
 
-            double? port,
+            int? port,
 
             string? privateKey,
